This brings the AVM implementation to the 'it compiles so it should work" state.  It's hooked up to nothing right now.  First will be a basic test harness for it, then...
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/BRANCHES/WCM-DEV2/root@2891 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/avm/AVMService.java
@@ -154,7 +154,7 @@ public interface AVMService
      * @param name The name of the Repository.
      * @return A Set of version IDs
      */
-    public Set<Integer> getRepositoryVersions(String name);
+    public Set<Long> getRepositoryVersions(String name);
     
     /**
      * Lookup a node by version ids and path.  Note that an AVMNode is essentially
